refactor: Payment external/kafka 패키지 정리
