+++ payment/src/main/java/com/truve/platform/payment/service/external/kafka/booking/BookingConsumer.java
@@ -1,10 +1,11 @@
-package com.truve.platform.payment.service.external.kafka;
+package com.truve.platform.payment.service.external.kafka.booking;
 
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.stereotype.Component;
 
 import com.truve.platform.common.support.JsonConverter;
+import com.truve.platform.payment.service.external.kafka.PaymentEventCommand;
 import com.truve.platform.payment.service.service.PaymentService;
 
 import lombok.RequiredArgsConstructor;
@@ -13,9 +14,9 @@ import lombok.extern.slf4j.Slf4j;
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class MembershipConsumer {
-	public static final String TOPIC = "membership.payment";
-	public static final String GROUP = "membership-payment-group";
+public class BookingConsumer {
+	public static final String TOPIC = "booking.payment";
+	public static final String GROUP = "booking-payment-group";
 
 	private final JsonConverter jsonConverter;
 	private final PaymentService paymentService;
